refactor: migrate sources.js to typescript
